Improve error message for ambiguous elements
Prior to this, when an element was in two disjoint master
sets, the program would abort with an error message saying
that no such set existed (which resulted from a check to
see if one set was a subset of another). Now the program
will abort with a message saying the element is a member of
multiple master sets, which is the underlying source of the
problem.