Support three new data files under Tablo
Adds support for files of additional parameters and endogenous
and exogenous variables. These support headers starting of the
form A### for endogenous variables, B### for exogenous variables,
and C### for parameters. The files would usually be used for new
variables and parameters added in the course of developing a new
version of a model.