Support -scalars option under -debug target.
Writes two additional files, <basename>_dec.csv and <basename>_use.csv
that show element-by-element declarations and usage of parameters and
variables. Useful for identifying scalars that are never used (as
happens if parameters or variables are defined over unnecessarily
large sets) or that are used in redundant equations (e.g., unexpectedly
occuring multiple times on the LHS of equation).